Remove typescript
The project was laready abusing typecasts to make certain things check out for the
typescript compiler, and with new features like static methods on Unions, more hacks
were used taht would make the library actually harder to work with in other typescript
projects.

The next step is to add es6 Symbol support, and it seems like typescript is just going
to cause even more problems. So it's gone, and it's all just es6 syntax now.