Change tests for new `Ok(result)` and `Some(maybe)` behaviour
Promise.resolve is like a cast: it promotes regular values to resolved
promises, or uses the promise passed to it. Results' API for Result.Ok() and
Maybe.Some() should match this.

The current behaviour is based off of a mis-reading of Promise.resolve.

These tests should validate the desired behaviour.