Implement union option instance .equals and .hashCode
hashCode currently always returns `42`, which is not efficient if used by a library like immutablejs, but is at least correct.

.equals does what it sounds like, checking that they're both union options, both union options from the same union, both the same member of the union, and finally checking
all the payload arguments (recursing, so nested structures are compared deeply).

The API "should" be enough for immutable.js to use as well, so union members can be embedded in immutabljs collections and be checked deeply from the top-level collections'
.equals method.

As a bonus, this commit fixes an issue in constructing the union option classes: previously, the constructor function's prototype was assigned to with a plain object in
order to set up the prototype methods. However, this caused each instance's `constructor` property to be `Object`. So now the methods are added by _assigning_ to
properties on the constructor function's prototype instead, and `.constructor` does what it should.

whew.