Expose options on the Union instance
This change eats up another key that now can't be an option member ('options'), in order to allow code to do some neat inspection of the union.

The motivating example: creating a copy of the union whose constructors are bound to do extra stuff. All I wanted was to iterate over the keys of the original. Now it's possible :)

It's recommended to use capital-case naming for union option members, so as long as that convention is followed, there is no new naming collision here.